Updated Styles and Inputs
Updated the styles on the page to work without the br tags.
Added an info button which tells you how to use the inputs. Needed to add a modal and it's styles for this.
Added an extra br to results so they're more distinguished.
Updated inputsto hopefully find a a fix for mobile keyboards.